Update bc_scraper.py
fixed an issue with file pathing.

Bash script running via cron uses the /home directory in linux, ergo previous pathing variables were rooted in /home and not the project directory.

using Path(__file__).parent.absolute() starts the pathing from the script's directory.

also removed mkdir, as these folders are in place and this is creating potential for unintentional directory creation if the script is ran out of place, as in this example. Not a necessary feature for now, will return to this if I want to make this app portable or plan on migrating it elsewhere.